chore: seperate auth apis into 2 end-points
