+++ apps/api-gateway/src/modules/auth/auth.controller.ts
@@ -1,12 +1,6 @@
 import {
   Body,
   Controller,
-  Delete,
-  ForbiddenException,
-  Get,
-  HttpCode,
-  HttpStatus,
-  Param,
   Patch,
   Post,
   Put,
@@ -15,30 +9,12 @@ import {
 } from '@nestjs/common';
 
 import { AuthService } from './auth.service';
-import {
-  AcceptFriendRequestDto,
-  BlockDto,
-  DeclineFriendRequestDto,
-  LogoutDto,
-  RefreshDto,
-  RegisterDto,
-  SendFriendRequestDto,
-  VerifyDto
-} from './dto';
-import { LoginDto } from './dto';
+import { LoginDto, RefreshDto, RegisterDto, VerifyDto } from './dto';
 import { EMAIL_REGEX } from '../../constants';
-import { JwtAtGuard, JwtRtGuard, JwtVtGuard } from '../../guards';
+import { JwtRtGuard, JwtVtGuard } from '../../guards';
 import { IRequestWithUser } from '../../types';
 
-import {
-  IGetBlockedResult,
-  IGetFriendRequestsResult,
-  IGetFriendsResult,
-  IGetMeResult,
-  IGetUserProfileResult,
-  ILoginResult,
-  IRefreshResult
-} from '@prj/types/api';
+import { ILoginResult, IRefreshResult } from '@prj/types/api';
 
 @Controller('auth')
 export class AuthController {
@@ -73,12 +49,6 @@ export class AuthController {
     return this.authService.verify({ id: req.user.sub });
   }
 
-  @UseGuards(JwtAtGuard)
-  @Get('me')
-  getMe(@Req() req: IRequestWithUser): Promise<IGetMeResult> {
-    return this.authService.getMe(req.user.sub);
-  }
-
   @UseGuards(JwtRtGuard)
   @Put('refresh')
   refresh(
@@ -89,160 +59,5 @@ export class AuthController {
       accountId: req.user.sub,
       refreshToken: body.refreshToken
     });
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Delete('logout')
-  logout(@Req() req: IRequestWithUser, @Body() body: LogoutDto) {
-    if (req.user.sub !== body.accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.logout({ accountId: body.accountId });
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Post(':id/friend-request')
-  sendFriendRequest(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Body() body: SendFriendRequestDto
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.sendFriendRequest(accountId, body);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Get(':id/friend-requests')
-  getPendingFriendRequest(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string
-  ): Promise<IGetFriendRequestsResult> {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.getPendingFriendRequest(accountId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @Patch(':id/friend-request/accept')
-  acceptFriendRequest(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Body() body: AcceptFriendRequestDto
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.acceptFriendRequest(accountId, body);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @Patch(':id/friend-request/decline')
-  declineFriendRequest(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Body() body: DeclineFriendRequestDto
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.declineFriendRequest(accountId, body);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @Delete(':id/friend-request/:target')
-  cancelFriendRequest(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Param('target') targetId: string
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.cancelFriendRequest(accountId, targetId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Get(':id/friends')
-  getFriends(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string
-  ): Promise<IGetFriendsResult> {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.getFriends(accountId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @Delete(':id/friend/:target')
-  removeFriend(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Param('target') targetId: string
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.removeFriend(accountId, targetId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Get(':id/profile')
-  async getUserProfile(
-    @Param('id') accountId: string,
-    @Req() req: IRequestWithUser
-  ): Promise<IGetUserProfileResult> {
-    return {
-      profile: await this.authService.getAccount({
-        id: accountId,
-        includeRelationshipWith: req.user.sub
-      })
-    };
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Get(':id/blocked')
-  async getBlockedUsers(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string
-  ): Promise<IGetBlockedResult> {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.getBlockedUsers(accountId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @Post(':id/block')
-  async blockUser(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Body() body: BlockDto
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.blockUser(accountId, body.targetId);
-  }
-
-  @UseGuards(JwtAtGuard)
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @Delete(':id/block/:target')
-  async unblockUser(
-    @Req() req: IRequestWithUser,
-    @Param('id') accountId: string,
-    @Param('target') targetId: string
-  ) {
-    if (req.user.sub !== accountId)
-      throw new ForbiddenException('Forbidden resource');
-
-    return this.authService.unblockUser(accountId, targetId);
   }
 }
