+++ apps/api-gateway/src/modules/auth/dto/index.ts
@@ -3,8 +3,3 @@ export * from './login.dto';
 export * from './refresh.dto';
 export * from './logout.dto';
 export * from './verify.dto';
-export * from './sendFriendRequest.dto';
-export * from './acceptFriendRequest.dto';
-export * from './declineFriendRequest.dto';
-export * from './block.dto';
-export * from './unblock.dto';
