+++ apps/api-gateway/src/modules/user/dto/acceptFriendRequest.dto.ts
