+++ apps/api-gateway/src/modules/user/dto/declineFriendRequest.dto.ts
