+++ apps/api-gateway/src/modules/user/dto/index.ts
@@ -1,0 +1,5 @@
+export * from './sendFriendRequest.dto';
+export * from './acceptFriendRequest.dto';
+export * from './declineFriendRequest.dto';
+export * from './block.dto';
+export * from './unblock.dto';
