+++ apps/api-gateway/src/modules/user/dto/sendFriendRequest.dto.ts
